+++ sdk/src/contexts/sei/index.ts
@@ -1,2 +1,2 @@
-export * from './context';
 export * from './contracts';
+export * from './context';
